+++ tourcompanydashboard/src/index.js
@@ -2,10 +2,16 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
+import AppErrorBoundary from './Components/ui/AppErrorBoundary';
+import { initializeMonitoring } from './monitoring';
+
+initializeMonitoring();
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>
+  <AppErrorBoundary>
+    <React.StrictMode>
+      <App />
+    </React.StrictMode>
+  </AppErrorBoundary>
 );